Fix -Wshorten-64-to-32 on 64bit part 2 d3d
- Use `checked_cast` for large buffer sizes, offsets.
- Use `static_cast` for values already gated by limits and
  frontend validation
  (e.g. texture width, uniform buffer size, attachment count)

Bug: 523358231
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/329135
Commit-Queue: Shrek Shao <shrekshao@google.com>

+++ src/dawn/native/d3d/ShaderUtils.cpp
@@ -158,10 +158,10 @@ ResultOrError<ComPtr<IDxcBlob>> CompileShaderDXC(const d3d::D3DBytecodeCompilati
     // pointers in this vector don't have static lifetime.
     std::vector<const wchar_t*> arguments = GetDXCArguments(entryPointW, r, hlslVersion);
     ComPtr<IDxcResult> result;
-    DAWN_TRY(CheckHRESULT(
-        r.dxcCompiler.UnsafeGetValue()->Compile(&dxcBuffer, arguments.data(), arguments.size(),
-                                                nullptr, IID_PPV_ARGS(&result)),
-        "DXC compile"));
+    DAWN_TRY(CheckHRESULT(r.dxcCompiler.UnsafeGetValue()->Compile(
+                              &dxcBuffer, arguments.data(), static_cast<uint32_t>(arguments.size()),
+                              nullptr, IID_PPV_ARGS(&result)),
+                          "DXC compile"));
 
     HRESULT hr;
     DAWN_TRY(CheckHRESULT(result->GetStatus(&hr), "DXC get status"));
